ディレクト分割
